input: stdin: Describe EoF behaviour, parser
Document how to use a parser with the stdin plugin to read arbitrary
messages, since the default behaviour is to assume a specific json input
format. This is very confusing for new users, since the plugin
just logs

    invalid record found, it's not a JSON map or array

when it sees inputs it doesn't expect, and the docs don't explain that
you can set a parser to handle other formats.

Document the stdin plugin's behaviour on end-of-file.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>