Update stackdriver.md
Adding this info about this issue because I lost a bunch of time trying to chase down intermittent restarts of my fluent-bit pods (v.2.2.0). 

Even if it defaults to 1 in newer versions, someone may override it, as was my case. 

Until the underlying issue is fixed, I think it's helpful to warn others about this.

Signed-off-by: Gerry Laracuente <122390399+gerry-tock@users.noreply.github.com>