Clarify tag expansion (#1366)
The `tail` docs don't make it clear how tag expansion of `*` actually works; I wasted a ton of time trying to configure filters due to not knowing that the slashes in the absolute path are replaced with dots. The only note about the replacement of the slashes with dots is in the docs for the kubernetes filter.

Could we also place a note about that behavior here? Or is this actually inaccurate and the kubernetes filter re-tags?

Signed-off-by: Jacob Page <85361835+jpage-godaddy@users.noreply.github.com>