Update kubernetes-events.md
This is a pretty massive change in the way the plugin works. Sticking this at the bottom in a foot note does not seem appropriate.

The decision to do this initially does make some sense, as 3.1 is not going to be the version that everyone is using. However, over time, this will come to be the dominant behavior, and that behavior should be made clear to the users.

Signed-off-by: Doug Whitfield <dwhitfield@perforce.com>