stackdriver: Refactor special fields docs
The special fields documentation as it existed was a bit of a mishmash
of old documentation for a previous logging agent solution and a lot of
additions over the years trying to match the old format. It ends up
reproducing a lot of information that is better stated by the existing
Cloud Logging docs, while missing crucial details about how the plugin
handled these cases. Certain fields were also missing.

This PR rewrites the documentation, doing more linking out to Google
Cloud Logging documentation in place of rewording the same information.
It also guides more generally on some of the edge cases for how the
special fields are handled, calling out the exceptions in a clearer way.
It also adds some fields that were previously missing from this
document.

Signed-off-by: braydonk <braydonk@google.com>