Alert NoOutputBytesProcessed should combine outputs (#640)
We were seeing issues where we got alerted due to some outputs not sending data for more than 15 minutes.
For example, this is expected from journald logs on worker nodes in a cluster, thus this adjustment to the alert.

We now aggregate output per pod and alert if no bytes are sent for 15 minutes.

Signed-off-by: Victor Sollerhed <victor.sollerhed@pagero.com>